# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Connect Chat Apps API",
        "description": "Dokumentasi API untuk Aplikasi Real-Time Chat & Voice Connect",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/servers": {
            "get": {
                "tags": [
                    "Servers"
                ],
                "summary": "Daftar Server yang Diikuti & Dimiliki User",
                "operationId": "e546310e6f6805e22f8fe3b88a74ded0",
                "responses": {
                    "200": {
                        "description": "Daftar server berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "servers": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servers"
                ],
                "summary": "Buat Server Baru",
                "operationId": "203cf2922c7414cfb2416b6a9a620d47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Komunitas Developer Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Ruang diskusi backend developer dan gamer"
                                    },
                                    "icon_url": {
                                        "type": "string",
                                        "example": "https://example.com/icon.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Server berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server berhasil dibuat."
                                        },
                                        "server": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi data gagal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/servers/{id}": {
            "get": {
                "tags": [
                    "Servers"
                ],
                "summary": "Detail Server beserta Daftar Channel dan Anggota",
                "operationId": "7b37b119418b683ad8ef51725fe3a111",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Server ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail server berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "server": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Bukan anggota server ini"
                    },
                    "404": {
                        "description": "Server tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Servers"
                ],
                "summary": "Update Informasi Server (Owner & Admin)",
                "operationId": "1e0fdcc9441e59fdd54f291480b2c54f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Server ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Komunitas Laravel & React"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Server resmi komunitas pengembang"
                                    },
                                    "icon_url": {
                                        "type": "string",
                                        "example": "https://example.com/new-icon.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Server berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server berhasil diperbarui."
                                        },
                                        "server": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak memiliki izin (Bukan Owner / Admin)"
                    },
                    "404": {
                        "description": "Server tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi data gagal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servers"
                ],
                "summary": "Hapus Server (Hanya Pemilik Server)",
                "operationId": "831eff57f53262fbc951d20e7572a2a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Server ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server berhasil dihapus."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak memiliki izin (Bukan pemilik server)"
                    },
                    "404": {
                        "description": "Server tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Sanctum kamu di sini",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}